move sort to different file
