+++ src/components/content/conditions/helpers/sorts.ts
@@ -1,5 +1,6 @@
 import { SortOption } from "@/components/core/sort-button/sort-button";
 import { ConditionModel } from "@/fhir/models";
+import { TimelineEventModel } from "@/fhir/models/timeline-event";
 
 export const defaultConditionSort: SortOption<ConditionModel> = {
   display: "Last Updated (New to Old)",
@@ -36,4 +37,17 @@ export const conditionSortOptions: SortOption<ConditionModel>[] = [
     sorts: [{ key: "recordedDate", dir: "asc", isDate: true }],
   },
   defaultConditionSort,
+];
+
+export const defaultTimelineSort: SortOption<TimelineEventModel> = {
+  display: "Date (New to Old)",
+  sorts: [{ key: "date", dir: "desc", isDate: true }],
+};
+
+export const timelineSortOptions: SortOption<TimelineEventModel>[] = [
+  defaultTimelineSort,
+  {
+    display: "Date (Old to New)",
+    sorts: [{ key: "date", dir: "asc", isDate: true }],
+  },
 ];
